feat(core/rust): cleaner backlight functions

+++ core/embed/rust/src/ui/model_tt/theme.rs
@@ -11,11 +11,11 @@ use crate::ui::{
 use super::component::{ButtonStyle, ButtonStyleSheet, LoaderStyle, LoaderStyleSheet};
 
 // Typical backlight values.
-pub const BACKLIGHT_NORMAL: i32 = 150;
-pub const BACKLIGHT_LOW: i32 = 45;
-pub const BACKLIGHT_DIM: i32 = 5;
-pub const BACKLIGHT_NONE: i32 = 2;
-pub const BACKLIGHT_MAX: i32 = 255;
+pub const BACKLIGHT_NORMAL: u16 = 150;
+pub const BACKLIGHT_LOW: u16 = 45;
+pub const BACKLIGHT_DIM: u16 = 5;
+pub const BACKLIGHT_NONE: u16 = 2;
+pub const BACKLIGHT_MAX: u16 = 255;
 
 // Color palette.
 pub const WHITE: Color = Color::rgb(0xFF, 0xFF, 0xFF);
